site: add report-http-installed code
This allows to send a HTTP(S) GET request
once the installation has finished.

This has to be called as late as possible,
i.e. after the final reboot and ideally
also after reorder_kernel has finished
since that usually corresponds to a usable system.

If enabled, the rc.firsttime cleanup will create a script for reporting,
deposited at /usr/local/bin/omir-report-http-installed.
This script will not be deleted.