site: reboot via /etc/rc.local
Found this on https://openbsd.amsterdam/setup.html

It's cleaner than hiding the reboot code in a rc.firsttime hook file.

However, I do not agree with their design decision as it might
result in a reboot loop should /etc be mounted readonly:

  if [ -r /etc/rc.local.forcereboot ]; then   # (true)
    rm -f /etc/rc.local.forcereboot           # could fail
    ...
    reboot                                    # possibly infinite reboot
  fi

This commit breaks the site.tgz temporarily.
The 99-cleanup-and-reboot hooks needs to be adjusted.